added some a new route for ticker (many stocks symbols functions)
